refactor(user): reemplazar controller y service con nuevas implementaciones
- Eliminar user.controllers.ts y user.services.ts antiguos

- Crear user.controller.ts y user.service.ts con todas las funciones async y roles admin

- Actualizar user.module.ts para reflejar los cambios
